[REFACTOR] : searchCadet 변수 변경 및 마이너 리팩터, 디렉토리 구조 변경 등
