+++ src/main/java/openproject/where42/api/mapper/SearchCadet.java
@@ -1,4 +1,4 @@
-package openproject.where42.api.dto;
+package openproject.where42.api.mapper;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.AllArgsConstructor;
@@ -17,8 +17,8 @@ import java.util.ArrayList;
 @NoArgsConstructor
 @AllArgsConstructor
 public class SearchCadet {
-    private String login;
-    private Image image;
+    private String name;
+    private String img;
     private String msg;
     private Locate locate;
     private int inOrOut;
@@ -26,9 +26,10 @@ public class SearchCadet {
     private String location;
     private boolean isMember;
 
+    // 멤버용
     public SearchCadet(Member member) {
-        this.login = member.getName();
-        this.image = new Image(member.getImg());
+        this.name = member.getName();
+        this.img = member.getImg();
         this.msg = member.getMsg();
         this.location = member.getLocation();
         this.isMember = true;
@@ -38,9 +39,10 @@ public class SearchCadet {
         }
     }
 
+    // 플래시 데이터
     public SearchCadet(FlashData flash) {
-        this.login = flash.getName();
-        this.image = new Image(flash.getImg());
+        this.name = flash.getName();
+        this.img = flash.getImg();
         this.location = flash.getLocation();
         if (Define.PARSED.equalsIgnoreCase(this.location)) {
             this.locate = flash.getLocate();
@@ -48,17 +50,22 @@ public class SearchCadet {
         }
     }
 
+    // 플래시 데이터에도 없는 경우 디비에 저장하지 않고 location null 셋팅
     public SearchCadet(String name, String img) {
-        this.login = name;
-        this.image = new Image(img);
-        this.location = Define.PARSED;
+        this.name = name;
+        if (this.img != null)
+            this.img = img;
+        else
+            this.img = "null";
         this.locate = new Locate(null, 0, 0, null);
         this.inOrOut = Define.NONE;
+        this.location = Define.PARSED;
     }
 
-    public SearchCadet(String name, String imgUrl, String msg, String spot) {
-        this.login = name;
-        this.image = new Image(imgUrl);
+    // 이스터 에그용
+    public SearchCadet(String name, String img, String msg, String spot) {
+        this.name = name;
+        this.img = img;
         this.msg = msg;
         this.locate = new Locate(null, 0, 0, spot);
         this.inOrOut = Define.IN;
@@ -74,6 +81,7 @@ public class SearchCadet {
         return where42s;
     }
 
+    // 파싱 시
     public void updateStatus(Locate locate, int inOrOut) {
         this.locate = locate;
         this.inOrOut = inOrOut;
