+++ src/main/java/openproject/where42/exception/customException/UnregisteredMemberException.java
@@ -1,7 +1,7 @@
 package openproject.where42.exception.customException;
 
 import lombok.Getter;
-import openproject.where42.api.dto.Seoul42;
+import openproject.where42.api.mapper.Seoul42;
 import openproject.where42.util.response.ResponseMsg;
 import openproject.where42.util.response.StatusCode;
